requirements.txt
apache-airflow>=2.7.3
pandas>=2.0.0
requests>=2.31.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
SQLAlchemy>=1.4.0
